fix: extract _launch_cook_session to _session_launch.py to satisfy app.py line limit
app.py was 773 lines, 18 over the 755-line architectural limit enforced by
test_app_py_under_line_limit. Moving _launch_cook_session to _session_launch.py
(where _run_interactive_session already lives) brings app.py to 737 lines.
