+++ src/autoskillit/cli/app.py
@@ -34,10 +34,9 @@ from autoskillit.cli._init_helpers import (
     _register_all,
 )
 from autoskillit.cli._prompts import _build_orchestrator_prompt, _get_ingredients_table
+from autoskillit.cli._session_launch import _launch_cook_session
 from autoskillit.core import (
-    NoResume,
     RecipeSource,
-    ResumeSpec,
     atomic_write,
     pkg_root,
     resume_spec_from_cli,
@@ -450,41 +449,6 @@ def recipes_render(name: str | None = None) -> None:
         sys.exit(1)
     diagram = load_recipe_diagram(name, _recipes_dir_for(match))
     print(diagram if diagram else f"No diagram. Run /render-recipe {name}")
-
-
-def _launch_cook_session(
-    system_prompt: str,
-    *,
-    initial_message: str | None = None,
-    extra_env: dict[str, str] | None = None,
-    resume_spec: ResumeSpec = NoResume(),
-    project_dir: Path | None = None,
-) -> None:
-    """Launch an interactive Claude Code cook session with reload loop support."""
-    from autoskillit.cli._session_launch import _run_interactive_session
-    from autoskillit.core import NamedResume
-
-    _max_reloads = 10
-    current_resume_spec = resume_spec
-    _current_initial_message = initial_message
-    seen_reload_ids: set[str] = set()
-    while True:
-        reload_id = _run_interactive_session(
-            system_prompt,
-            initial_message=_current_initial_message,
-            extra_env=extra_env,
-            resume_spec=current_resume_spec,
-            project_dir=project_dir,
-        )
-        if reload_id is None:
-            break
-        if len(seen_reload_ids) >= _max_reloads:
-            raise SystemExit(f"Too many reloads ({_max_reloads} max). Check for infinite loop.")
-        if reload_id in seen_reload_ids:
-            raise SystemExit(f"Repeated reload_id {reload_id!r} — aborting.")
-        seen_reload_ids.add(reload_id)
-        current_resume_spec = NamedResume(session_id=reload_id)
-        _current_initial_message = None
 
 
 def _get_subsets_needed(recipe: Recipe, disabled_subsets: frozenset[str]) -> frozenset[str]:
